requirements.txt
Flask==3.0.3
torch==2.4.1
Pillow==10.4.0
numpy==1.24.4
opencv-python==4.8.0.74
ultralytics==8.2.93
